extract the base template to a separate file
